spotify-recommendations/requirements.txt
hdbscan
spotipy
django
funcy
matplotlib
numpy
seaborn
pandas
django-picklefield
scikit_learn
social-auth-app-django
gunicorn
whitenoise
